+++ src/commands/storageAccountCreateSimple.js
@@ -15,9 +15,37 @@ exports.createCommand = function createCommand(state) {
             state.resourceGroupToUse = state.selectedStorageAccount + 'Resources';
 
             ux.ifStorageAccountNameIsAvailable(state).then(() => {
-                ux.createResourceGroup(state, () => {
-                    ux.createStorageAccount(state);
+                vscode.window.showQuickPick([
+                    constants.optionExistingRg,
+                    constants.optionNewRg
+                ]).then(selected => {
+                    if (selected == constants.optionExistingRg) {
+                        ux
+                            .getResourceGroups(state)
+                            .then(function () {
+                                vscode.window.showQuickPick(state.resourceGroupList)
+                                    .then(function (selectedRg) {
+                                        if (!selectedRg) return;
+                                        state.resourceGroupToUse = selectedRg;
+                                        ux.createStorageAccount(state);
+                                    });
+                            })
+                            .catch(function (err) {
+                                vscode.window.showErrorMessage(err);
+                            });
+                    }
+                    else if (selected == constants.optionNewRg) {
+                        vscode.window.showInputBox({
+                            prompt: constants.promptNewRgName
+                        }).then(function (newResourceGroupName) {
+                            state.resourceGroupToUse = newResourceGroupName;
+                            ux.createResourceGroup(state, () => {
+                                ux.createStorageAccount(state);
+                            });
+                        });
+                    }
                 });
+
             }).catch((message) => {
                 vscode.window.showErrorMessage(message);
             });